requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
idna==3.4
loguru==0.6.0
numpy==1.24.2
pandas==2.0.0
pinecone-client==2.2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
